Added divs for html files
Divs and the class attributes were added to the headings in the main tag
of Index.html. This is so I can change the headings to black. I also
added a div and class attributes to the home.png so that when I add
other images, the css for the other images won't change home.png.
Padding was added to the previous divs (About, Hackathon Projects,
Projects, Experience ((Index.html files)). For contact.html, divs and
class attributes were added as well to separate the links. The Margin
for the social links was increased. These changes were made because I
wanted to have more control and be able to access the other tags with
css more.