Debugging and image fixes
Fixed some of the issues in both html files and css files with an html validator. The title tag for home was replaced and I put the home tag in the nav bar. I put the main tag inside the body tag along with sizing the favicon better.  I have also added a line break for line 29 to make it more readable.  I removed the div tag for home.png as I don't want too many div tags.  I made these changes as I started to debug my html and make it better. I will find a way to put space in between the Home and Contact me in the nav bar. I will also need to separate the projects into different columns as well.