Fixing open tags
Index.html changes:-Added closing tags for <br> on line 34 and removed
the extra <head> tag on line 13 or 14

Contact.html changes: -Removed comment for line 26 or 27 as I no longer
need it.

I made these changes so that the website would work better and look the
way I want it to.