Added links and another photo
Contact.html: Added a link for LinkedIn. I don't think it is hotlinking.

Index.html: Added a link to the Button of Happiness website and I
changed the status of the matrix calculator as it is done. It
currently doesn't work. Also added the image of the island of Hawaii
that I took. I used sirv to optimize the image somewhat.

Styles.css: Added more onto the img with flexbox but I'm not too sure
there is a difference. I will try to get the photo of Hawaii into the
center but it'll take a while.

Hawaii.JPG: This is the unoptimized photo and I will most likely remove
the sirv and set the image myself.

I made these changes as I feel that the website is quite empty without a
photo and I want to have more links.