Added visted for nav bar
The color for when you visit a certain link is #E1951E. It looks like a
golden yellow to me. This is change affects both files styles.css and
contact.css.
Additional changes for Styles.css: There is now a border around the
photo in the center.
I made these changes as I thought it would be more interesting to get
the visited pseudo-class. I think I had this change before and it seems
good to have it back.