Starting to add toggle bar
First of all, this is an incorrect usage of the details tag and summary
tag. I looked up the mdn documentation and thought it worked. I don't
recommended that you do this at all. However, this is one of the first steps I'm taking to create a toggle for the nav bar as I think it would look more clean if I did so. It opens up the nav bar like a toggle and I'll replace it soon.