Changed display of the project tabs
In index.html, there aren't any changes besides cleaning up the html I
think. In styles/styles.css, I changed the display to inline-block in
Projects p. On line 10, justify content is now space-between instead of
space-around. I made these changes as I wanted to make the projects not
centered and instead, it is on a row.