Starting to add Hamburger icon.
I am starting to do the hambruger icon that I tried to do last time but
couldn't. So I searched up for a tutorial online and look at how it is
done. In index.html, I added the button with a class and an Id as well
as the nav bar getting an id. In styles.css, in lines 114 -126 which are
commented off, I plan to add the menu to make the nav bar more
accessible to mobile devices.I also added cursor: pointer for lines
14-16. I made these changes so that mobile users can access the website
better and not have to deal with a nav bar that is split and compressed.