Added css file for font awesome
For Index.html, I added the css file for font awesome and the javascript
file was moved down to the bottom of the body tag at line 66. In
styles.css, I removed a duplicate element for nav or something along the
lines of that. I made these changes so that the font awesome icon can
show.