Starting to add SEO
I added a description for meta data in index.html and added
rel="noreferrer" for links in both index.html and contact.html. I also
commented out the js file. I will start to optimize photos as that is
what is taking most of the loading time for this website. I am putting
the mobile view on hold for now and doing this instead.