Nav bar changes
In index.html, I have made it so that wheen you click o one of the
things on the nav bar such as about, it scrolls down to there. I also
changed it so that mobile users can now zoom in now. These changes are
quite small and I'm trying to work towards getting mobile loooking
better.